adds rate

+++ db/migrate/20150418000855_create_debts.rb
@@ -2,8 +2,8 @@ class CreateDebts < ActiveRecord::Migration
   def change
     create_table :debts do |t|
     	t.string :type
-    	t.integer :amount
-    	t.integer :monthly_payment
+    	t.decimal :amount
+    	t.decimal :monthly_payment
     	t.decimal :rate
     	t.string :act_primary
     	t.string :act_secondary
